requirements.txt
annotated-types==0.7.0
anyio==4.13.0
atproto==0.0.65
certifi==2026.4.22
cffi==2.0.0
click==8.3.3
cryptography==46.0.7
dnspython==2.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.14
libipld==3.3.2
pycparser==3.0
pydantic==2.13.4
pydantic_core==2.46.4
python-dotenv==1.2.2
typing-inspection==0.4.2
typing_extensions==4.15.0
websockets==15.0.1
